avocado: Add avocado.utils.stacktrace for stack trace related utilities
Refactor the test code and create a library containing
stack trace related utilities for avocado. They'll be
useful in a later patch, that improves visibility of
errors in the test loading stage.

Changes from v1:
 * Changed module name from tb to stacktrace, per rmoura's suggestion.
